feat: 플랫폼 별 대응 코드 추가
