+++ src/screens/HomeScreen.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
-import { Dimensions, Linking, Platform, StatusBar, View } from 'react-native';
+import { Linking, Platform, View } from 'react-native';
 import { WebView, WebViewNavigation } from 'react-native-webview';
 
+import { useAndroidSafeArea } from '~/hooks/useAndroidSafeArea';
 import theme from '~/styles/theme';
 
 const uri = 'https://app.ygtang.kr/';
 
 export default function HomeScreen() {
+  const { androidSafeAreaInjectScript } = useAndroidSafeArea();
+
   const handleExternalLinks = (event: WebViewNavigation) => {
     const isExternalLink = Platform.OS === 'ios' ? event.navigationType === 'click' : true;
     if (isExternalLink) {
@@ -20,31 +23,16 @@ export default function HomeScreen() {
     return true;
   };
 
-  const INJECTED_JAVASCRIPT = `(function() {
-    const style = document.createElement('style');
-    style.innerHTML = \`
-      .safeAreaTop {
-        height: ${StatusBar.currentHeight}px !important;
-      }
-      .safeAreaBottom {
-        height: ${Dimensions.get('screen').height - Dimensions.get('window').height}px !important;
-      }
-    \`;
-    document.head.appendChild(style);
-  })();`;
-
   return (
     <View style={{ flex: 1, backgroundColor: theme.color.background }}>
-      <StatusBar
-        translucent
-        backgroundColor="transparent"
-        barStyle={'dark-content'}
-        hidden={false}
-      />
       <WebView
         source={{ uri }}
         bounces={false}
-        injectedJavaScript={INJECTED_JAVASCRIPT}
+        injectedJavaScript={
+          `(function(){
+            ${[androidSafeAreaInjectScript].join('\n')}
+          })();` as string
+        }
         applicationNameForUserAgent={'YgtangApp/1.0'}
         domStorageEnabled
         onNavigationStateChange={handleExternalLinks}
